Done by API Task

+++ WebApiProject/ErrorLog/LogNLog.cs
@@ -16,18 +16,39 @@ namespace WebApiProject.ErrorLog
         }
         public List<LoggingError> GetLog()
         {
+            if (!_context.MyLog.Any())
+            {
+               //      Console.WriteLine("No Content Available in the Table");
+                return _context.MyLog.ToList();
+
+            }
             return _context.MyLog.ToList();
         }
-        public void Delete(string entity, string type)
+        public string[] Delete(string entity, string type)
+        { 
+            //  bool b = _context.MyLog.Where(r => r.Created.Contains(ee) && r.Type.Contains(type));
+
+              //var _dateTime = _context.MyLog.Where(r => r.Created.Contains(ee) && r.Type.Contains(type) );
+            if (!DataExists(entity, type))
+            {
+                return new string[] { "Content Not Found" };
+
+            }
+            else
+            {
+                string ee = entity.Replace('-', '/');
+                var _dateTime = _context.MyLog.Where(r => r.Created.Contains(ee) && r.Type.Contains(type));
+                _context.MyLog.RemoveRange(_dateTime);
+                _context.SaveChanges();
+                return new string[] { "Deleted" };
+            }
+
+        }
+      public bool DataExists(string entity, string type)
         {
             string ee = entity.Replace('-', '/');
-           // bool b = _context.MyLog.Where(r => r.Created.Contains(ee)) && _context.MyLog.Where(r => r.Type.Contains(type));
-         
-           var _dateTime = _context.MyLog.Where(r => r.Created.Contains(ee) && r.Type.Contains(type) );     
-           _context.MyLog.RemoveRange(_dateTime);
-           _context.SaveChanges();  
+            return _context.MyLog.Any(r => r.Created.Contains(ee) && r.Type.Contains(type));
         }
-        
         public void SetLog(string Message)
         {
             LoggingError logging = new LoggingError
@@ -60,6 +81,11 @@ namespace WebApiProject.ErrorLog
             };
             _context.MyLog.Add(logging);
             _context.SaveChanges();
+        }
+
+        string ILog.Delete(string entity, string type)
+        {
+            throw new NotImplementedException();
         }
     }
 }
